Renamed CalcByArrival to CalcDay, fixed midday calculation and time picker functionality
