--- app/src/main/java/com/example/hours/ui/calcByArrival/CalcByArrivalViewModel.java
@@ -1,7 +1,0 @@
-package com.example.hours.ui.calcByArrival;
-
-import androidx.lifecycle.ViewModel;
-
-public class CalcByArrivalViewModel extends ViewModel {
-    // TODO: Implement the ViewModel
-}
