+++ app/src/main/java/com/example/hours/ui/calcDay/CalcDayViewModel.java
@@ -1,0 +1,7 @@
+package com.example.hours.ui.calcDay;
+
+import androidx.lifecycle.ViewModel;
+
+public class CalcDayViewModel extends ViewModel {
+    // TODO: Implement the ViewModel
+}
