+++ app/src/main/res/layout/fragment_calc_day.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@drawable/background_app"
-    tools:context=".ui.calcByArrival.CalcByArrivalFragment">
+    tools:context=".ui.calcDay.CalcDayFragment">
 
     <ScrollView
         android:layout_width="match_parent"
