+++ app/src/test/java/com/example/hours/HoursManagerTest.java
@@ -23,7 +23,7 @@ public class HoursManagerTest extends TestCase {
     public void testGetInfoByArrivalTime1() {
         HoursInfo info = new HoursInfo();
         info.mArrivalTime = new Timestamp(7, 30);
-        HoursInfo resHoursInfo = HoursManager.getInstance().GetInfoByArrivalTime(info);
+        HoursInfo resHoursInfo = HoursManager.getInstance().CalcDay(info);
         HoursInfo expHoursInfo = new HoursInfo();
         expHoursInfo.mArrivalTime = new Timestamp(7, 30);
         expHoursInfo.mHalfDay = new Timestamp(11, 42);
@@ -40,7 +40,7 @@ public class HoursManagerTest extends TestCase {
     public void testGetInfoByArrivalTime2() {
         HoursInfo info = new HoursInfo();
         info.mArrivalTime = new Timestamp(7, 49);
-        HoursInfo resHoursInfo = HoursManager.getInstance().GetInfoByArrivalTime(info);
+        HoursInfo resHoursInfo = HoursManager.getInstance().CalcDay(info);
         HoursInfo expHoursInfo = new HoursInfo();
         expHoursInfo.mArrivalTime = new Timestamp(7, 49);
         expHoursInfo.mHalfDay = new Timestamp(12, 01);
